moved matchers to libraries
fixes #11
