+++ spec/spec_helper.rb
@@ -1,3 +1,2 @@
 require 'chefspec'
 require 'chefspec/berkshelf'
-require 'support/matchers'
